+++ lib/src/base_repository.dart
@@ -1,51 +1,55 @@
 import 'dart:async';
 
-import 'package:meta/meta.dart';
 import 'package:repository/src/domain/entities/data_source.dart';
 import 'package:repository/src/domain/entities/repository_state.dart';
 import 'package:repository/src/infra/repository_cache_storage.dart';
 import 'package:repository/src/infra/repository_fiber.dart';
 import 'package:repository/src/infra/repository_logger.dart';
 import 'package:repository/src/repositories/http_repository.dart';
+import 'package:meta/meta.dart';
 import 'package:retry/retry.dart';
 import 'package:rxdart/rxdart.dart';
 
-/// A [Repository] that can be backpropagated to a remote source.
+/// A [BaseRepository] that can be backpropagated to a remote source.
 /// It is useful when you want to save data to a remote source
 /// though a repository.
 /// For example, you can use this mixin to save data to a remote API
 /// when a user updates a profile.
-mixin PropagatingRepositoryMixin<Data> on Repository<Data> {
+mixin MutatorRepositoryMixin<Data> on BaseRepository<Data> {
   /// Propagates data to a remote source and updates the stream.
-  Future<void> propagate(Data data);
+  Future<void> mutate(Data data);
 }
 
-/// A [Repository] is a class that holds data and provides a stream.
+/// A [BaseRepository] is a class that holds data and provides a stream.
 /// It can be used to fetch data from a remote source, cache it, and provide a
 /// stream of that data.
-abstract class Repository<Data> {
+abstract class BaseRepository<Data> {
   /// If [resolveOnCreate] is true, the repository will resolve itself on
   /// creation.
   /// If [autoRefreshInterval] is not null, the repository will refresh itself
   /// every [autoRefreshInterval].
-  Repository({
+  BaseRepository({
     this.autoRefreshInterval,
     bool resolveOnCreate = true,
-  }) {
+    List<Repository<dynamic>>? dependencies,
+  }) : dependencies = dependencies ?? <Repository<dynamic>>[] {
     track();
 
-    hydrate(refreshAfter: resolveOnCreate);
+    hydratate(refreshAfter: resolveOnCreate);
 
     if (autoRefreshInterval != null) {
-      timer = Timer.periodic(
-        autoRefreshInterval!,
-        (_) => refresh(),
-      );
+      timer = Timer.periodic(autoRefreshInterval!, (_) {
+        if (_controller.hasListener) {
+          refresh();
+        }
+      });
     }
+
+    _listenToDependencies();
   }
 
   /// {@macro http_repository}
-  factory Repository.http({
+  factory BaseRepository.http({
     required Uri endpoint,
     Data Function(String json)? fromJson,
     FutureOr<bool> Function(Exception exception)? shouldRetryCondition,
@@ -54,7 +58,7 @@ abstract class Repository<Data> {
     bool resolveOnCreate = true,
     String? name,
   }) {
-    return HttpRepository<Data>(
+    return Repository<Data>(
       name: name,
       endpoint: endpoint,
       fromJson: fromJson,
@@ -65,19 +69,46 @@ abstract class Repository<Data> {
     );
   }
 
+  void addDependency(Repository<dynamic> dependency) {
+    _unlistenToDependencies();
+    dependencies.add(dependency);
+    _listenToDependencies();
+  }
+
+  final _subscriptions = <StreamSubscription<RepositoryState<dynamic>>>[];
+
+  void _listenToDependencies() {
+    _subscriptions.addAll(
+      dependencies.map(
+        (dependency) => dependency.stream.listen((state) {
+          if (state is RepositoryStateReady) {
+            refresh();
+          }
+        }),
+      ),
+    );
+  }
+
+  void _unlistenToDependencies() {
+    for (final subscription in _subscriptions) {
+      subscription.cancel();
+    }
+    _subscriptions.clear();
+  }
+
   /// List of all repositories in memory. It's useful for debugging.
-  static final List<WeakReference<Repository<dynamic>>> repositories = [];
+  static final List<WeakReference<BaseRepository<dynamic>>> repositories = [];
 
   /// Add the repository to the list of all repositories in memory if it's not
   /// already in the list.
   @protected
   void track() {
-    final alreadyTracked = Repository.repositories.any(
+    final alreadyTracked = BaseRepository.repositories.any(
       (ref) => ref.target?.key == key,
     );
 
     if (!alreadyTracked) {
-      Repository.repositories.add(WeakReference(this));
+      BaseRepository.repositories.add(WeakReference(this));
     }
   }
 
@@ -98,7 +129,7 @@ abstract class Repository<Data> {
   /// Get the current data of the repository
   /// if it's already resolved, otherwise resolve it.
   /// This method will not refresh the repository if it's already resolved.
-  Future<Data> currentValueOrResolve() async {
+  Future<Data?> currentValueOrResolve() async {
     return currentValue ?? await refresh();
   }
 
@@ -121,10 +152,7 @@ abstract class Repository<Data> {
   /// Getter for the last value of the stream.
   /// Returns null if the stream is empty.
   Data? get currentValue {
-    return currentState.map(
-      empty: (_) => null,
-      ready: (state) => state.data,
-    );
+    return currentState.map(empty: (_) => null, ready: (state) => state.data);
   }
 
   /// Returns the current state of the repository.
@@ -144,11 +172,14 @@ abstract class Repository<Data> {
 
   /// The `Fiber` is used to avoid multiple refreshes at the same time.
   @protected
-  final refreshFiber = RepositoryFiber<Data>();
+  final refreshFiber = RepositoryFiber<Data?>();
 
   /// The `Fiber` is used to avoid multiple hydratations at the same time.
   @protected
-  final hydratationFiber = RepositoryFiber<Data?>();
+  final _hydratationFiber = RepositoryFiber<Data?>();
+
+  @protected
+  final Completer<Data?> hydratationCompleter = Completer<Data?>();
 
   /// Disposes the repository. You should call this method when you're done
   /// using the repository.
@@ -156,6 +187,7 @@ abstract class Repository<Data> {
   /// You should not use the repository after calling this method.
   void dispose() {
     timer?.cancel();
+    _unlistenToDependencies();
     _controller.close();
   }
 
@@ -167,14 +199,18 @@ abstract class Repository<Data> {
   /// Gets the data from the cache, if it exists, and emits it to the stream.
   @visibleForTesting
   @protected
-  Future<Data?> hydrate({bool refreshAfter = true}) async {
-    return hydratationFiber.run(() async {
+  Future<Data?> hydratate({bool refreshAfter = true}) async {
+    return _hydratationFiber.run(name: name, () async {
       final stopwatch = Stopwatch()..start();
       try {
         final cachedDataString = await storage.read(key: key);
 
         if (cachedDataString != null) {
-          return await _emitRawData(cachedDataString);
+          final data = await _emitRawData(cachedDataString);
+          if (!hydratationCompleter.isCompleted) {
+            hydratationCompleter.complete(data);
+          }
+          return data;
         }
       } on FormatException catch (e) {
         logger.call(
@@ -190,6 +226,10 @@ abstract class Repository<Data> {
           'hydrated in ${stopwatch.elapsedMilliseconds}ms',
         );
 
+        if (!hydratationCompleter.isCompleted) {
+          hydratationCompleter.complete(null);
+        }
+
         if (refreshAfter) {
           await refresh();
         }
@@ -203,10 +243,7 @@ abstract class Repository<Data> {
     RepositoryDatasource datasource = RepositoryDatasource.local,
   }) async {
     final data = fromJson(rawData);
-    await emit(
-      data: data,
-      datasource: datasource,
-    );
+    await emit(data: data, datasource: datasource);
 
     // We do not need to persist if it comes from the cache or
     // if the data is optimistic.
@@ -218,11 +255,14 @@ abstract class Repository<Data> {
   }
 
   /// Refreshes the repository from remote datasource.
-  Future<Data> refresh() async {
+  Future<Data?> refresh() async {
     return retry(
       // Run the refresh in a fiber to avoid multiple refreshes at the same time
-      () => refreshFiber.run(_refresh),
+      () => refreshFiber.run(name: name, _refresh),
       retryIf: shouldRetry,
+      onRetry: (exception) {
+        logger('Repository($name): Retrying refresh...');
+      },
     );
   }
 
@@ -230,21 +270,24 @@ abstract class Repository<Data> {
   @protected
   FutureOr<bool> shouldRetry(Exception exception) => true;
 
-  Future<Data> _refresh() async {
+  Future<Data?> _refresh() async {
     // Save the current time to calculate the time it took to refresh the
     final before = DateTime.now();
     // Resolve the data from the remote source
     final rawData = await resolve();
     // Decodes the raw data to the data that will be used in the stream
     // Emit the data to the stream and persist
-    final data = await _emitRawData(
-      rawData,
-      datasource: RepositoryDatasource.remote,
-    );
+    Data? data;
+    if (rawData != null) {
+      data = await _emitRawData(
+        rawData,
+        datasource: RepositoryDatasource.remote,
+      );
+    }
     // Log the time it took to refresh
     final after = DateTime.now();
     final timeSpent = after.difference(before);
-    Repository.logger(
+    BaseRepository.logger(
       'Repository($name): refreshed in'
       ' ${timeSpent.inMilliseconds}ms',
     );
@@ -263,17 +306,7 @@ abstract class Repository<Data> {
     // Call the resolver function to get the new data.
     final newData = resolver.call(currentValue);
     // Add the new data to the repository without refreshing yet.
-    await emit(
-      data: newData,
-      datasource: RepositoryDatasource.optimistic,
-    );
-    // It's just 'cause `this` is a getter, so the 'if' below will not work
-    // if we don't use it as a local variable.
-    final self = this;
-
-    if (self is PropagatingRepositoryMixin<Data>) {
-      await self.propagate(newData);
-    }
+    await emit(data: newData, datasource: RepositoryDatasource.optimistic);
 
     // Refresh the repository.
     await refresh();
@@ -286,12 +319,7 @@ abstract class Repository<Data> {
     RepositoryDatasource datasource = RepositoryDatasource.local,
   }) async {
     logger('Emitting data to repository $name: $data');
-    _controller.add(
-      RepositoryState.ready(
-        data: data,
-        source: datasource,
-      ),
-    );
+    _controller.add(RepositoryState.ready(data: data, source: datasource));
   }
 
   /// Clears the cache and emits an empty state to the repository stream.
@@ -306,7 +334,7 @@ abstract class Repository<Data> {
   /// The returned string will be saved in the cache and decoded using
   /// [fromJson].
   @protected
-  Future<String> resolve();
+  Future<String?> resolve();
 
   /// Transforms the raw data from the remote source to the data that will be
   /// used in the stream.
@@ -333,4 +361,9 @@ abstract class Repository<Data> {
   /// The data will be refreshed every [autoRefreshInterval].
   /// The data will be refreshed when [refresh] is called.
   late final Stream<RepositoryState<Data>> stream = _controller.stream;
+  late final Stream<Data?> dataStream = stream.map(
+    (state) => state.map(ready: (state) => state.data, empty: (_) => null),
+  );
+
+  final List<Repository<dynamic>> dependencies;
 }