+++ lib/src/domain/entities/data_source.dart
@@ -7,6 +7,6 @@ enum RepositoryDatasource {
   local,
 
   /// The data was injected by the `update` method
-  /// of `PropagatingRepositoryMixin`.
+  /// of `MutatorRepositoryMixin`.
   optimistic,
 }