+++ lib/src/infra/repository_http_client.dart
@@ -1,4 +1,7 @@
-import 'dart:io';
+import 'dart:io' show HttpClientRequest;
+
+import 'package:equatable/equatable.dart';
+import 'package:meta/meta.dart';
 
 /// A type alias for bearer token String.
 /// Notice that this is just the token, without the `Bearer` prefix.
@@ -11,21 +14,24 @@ typedef BearerToken = String;
 /// {@endtemplate}
 abstract class RepositoryHttpClient {
   /// {@macro repository_http_client}
-  const RepositoryHttpClient();
+  const RepositoryHttpClient({this.mocks});
+
+  final Map<RepositoryHttpMockedRequest, RepositoryHttpResponse>? mocks;
 
   /// Makes a HTTP `get` request using [HttpClientRequest].
-  Future<RepositoryHttpResponse> get({
-    required RepositoryHttpRequest request,
-  });
+  Future<RepositoryHttpResponse> call({required RepositoryHttpRequest request});
 
-  /// It's used to get the token to be used in the
-  /// `Authorization` header. It's also a static variable,
-  /// so you can set it once and it will be used in all requests.
-  static TokenBuilder? tokenBuilder;
+  /// Handle mocked requests if needed.
+  @protected
+  RepositoryHttpResponse? findMock(RepositoryHttpRequest request) {
+    final mock = RepositoryHttpMockedRequest(
+      url: request.url,
+      method: request.method,
+    );
+
+    return mocks?[mock];
+  }
 }
-
-/// A type alias for a function that returns a [BearerToken].
-typedef TokenBuilder = Future<BearerToken?> Function();
 
 /// {@template repository_http_response}
 /// A class that represents the response from a HTTP request.
@@ -56,6 +62,9 @@ class RepositoryHttpResponse {
   }
 }
 
+/// A type alias for the HTTP method.
+enum RepositoryHttpMethod { get, post, put, delete, patch }
+
 /// {@template repository_http_request}
 /// This is used by [RepositoryHttpClient] to make HTTP requests.
 /// You can use this class to create your own HTTP client,
@@ -66,17 +75,38 @@ class RepositoryHttpRequest {
   /// {@macro repository_http_request}
   const RepositoryHttpRequest({
     required this.url,
-    Map<String, String>? headers,
-  }) : headers = headers ?? const {};
+    this.method = RepositoryHttpMethod.get,
+    this.body = const {},
+    this.headers = const {},
+  });
 
   /// The url of the request.
   final Uri url;
 
+  /// The method of the request.
+  final RepositoryHttpMethod method;
+
   /// The headers of the request.
   final Map<String, String> headers;
+
+  /// The body of the request.
+  final Map<String, dynamic>? body;
 
   @override
   String toString() {
     return 'RepositoryHttpRequest{url: $url, headers: $headers}';
   }
+}
+
+class RepositoryHttpMockedRequest extends Equatable {
+  const RepositoryHttpMockedRequest({required this.url, required this.method});
+
+  /// The url of the request.
+  final Uri url;
+
+  /// The method of the request.
+  final RepositoryHttpMethod method;
+
+  @override
+  List<Object?> get props => [url, method];
 }