+++ lib/src/repositories/http_repository.dart
@@ -1,60 +1,190 @@
 import 'dart:async';
+import 'dart:io' show SocketException;
 
-import 'package:repository/src/repositories/custom_http_repository.dart';
+import 'package:repository/src/base_repository.dart';
+import 'package:repository/src/domain/exceptions/network_unavailable_exception.dart';
+import 'package:repository/src/domain/exceptions/unexpected_status_code_exception.dart';
+import 'package:repository/src/external/platform_http_client.dart';
+import 'package:repository/src/infra/repository_http_client.dart';
+import 'package:repository/src/infra/repository_logger.dart';
 
 /// {@template http_repository}
-/// A `Repository` that fetches data from an endpoint. It is a wrapper around
-/// [CustomHttpRepository] that uses the default `RepositoryHttpClient`.
-/// It is recommended to use [CustomHttpRepository] instead of this class if you
-/// need to customize the request.
+/// A `Repository` that fetches data from an HTTP endpoint.
+/// Can be used both directly (instanciation) and by inheritance.
+///
+/// Usage examples:
+///
+/// Direct usage:
+/// ```dart
+/// final repo = Repository<MyData>(
+///   endpoint: Uri.parse('https://api.example.com/data'),
+///   fromJson: (json) => MyData.fromJson(json),
+/// );
+/// ```
+///
+/// By inheritance:
+/// ```dart
+/// class MyRepository extends Repository<MyData> {
+///   MyRepository() : super(endpoint: Uri.parse('https://api.example.com/data'));
+///
+///   @override
+///   MyData fromJson(String json) => MyData.fromJson(json);
+/// }
+/// ```
 /// {@endtemplate}
-class HttpRepository<Data> extends CustomHttpRepository<Data> {
-  /// Creates a [HttpRepository] that fetches data from an endpoint.
-  /// The [endpoint] is the only required parameter.
-  /// The [fromJson] function is optional
-  ///  and defaults to returning the json as is.
-  /// The `autoRefreshInterval` is optional and defaults to null.
-  /// The `resolveOnCreate` is optional and defaults to true.
-  /// If `autoRefreshInterval` is not null, the repository will automatically
-  HttpRepository({
+class Repository<Data> extends BaseRepository<Data>
+    with MutatorRepositoryMixin<Data> {
+  /// Creates an [Repository] that fetches data from an endpoint.
+  ///
+  /// The [endpoint] is the only required parameter when used directly.
+  /// The [fromJson] function is optional and defaults to returning the json as is.
+  /// The [autoRefreshInterval] is optional and defaults to null.
+  /// The [resolveOnCreate] is optional and defaults to true.
+  /// If [autoRefreshInterval] is not null, the repository will automatically
+  /// refresh at the specified interval.
+  Repository({
     required this.endpoint,
-    Data Function(String json)? fromJson,
-    FutureOr<bool> Function(Exception exception)? shouldRetryCondition,
+    this._fromJson,
+    this._shouldRetryCondition,
     super.resolveOnCreate,
     super.autoRefreshInterval,
-    super.tag,
-    String? name,
-  })  : _fromJson = fromJson ?? ((json) => json as Data),
-        _shouldRetryCondition = shouldRetryCondition,
-        _name = name,
-        super();
+    this.tag,
+    this._name,
+    this._mutate,
+    this.method = RepositoryHttpMethod.get,
+    super.dependencies,
+  }) : super();
 
+  final RepositoryHttpMethod method;
   final String? _name;
+  final Data Function(String json)? _fromJson;
+  final FutureOr<bool> Function(Exception exception)? _shouldRetryCondition;
+  final Future<void> Function(Data data)? _mutate;
 
-  @override
-  String get name => _name ?? super.name;
-
-  /// The endpoint to fetch the data from.
+  /// The endpoint to fetch data from.
   /// This is the only required parameter.
-  @override
   final Uri endpoint;
 
-  /// Private field that holds the [fromJson] callback.
-  final Data Function(String json) _fromJson;
+  /// The tag of the repository.
+  /// This is used to identify the repository in the cache.
+  /// If the tag is not null, the repository will use the tag to
+  /// create unique keys for the cache.
+  /// A common use case for this is when you want to cache for different
+  /// users. In this case, you can use the user id (e.g. e-mail) as the tag.
+  @override
+  final String? tag;
 
-  /// Private field that holds the [onSocketException] callback.
-  final FutureOr<bool> Function(Exception exception)? _shouldRetryCondition;
+  /// The http client used to fetch data from the endpoint.
+  /// This is a monostate, so it will be shared across all instances of
+  /// [Repository].
+  static final RepositoryHttpClient _client = createPlatformHttpClient();
+  RepositoryHttpClient get client => _client;
+
+  @override
+  String get name => _name ?? endpoint.path.split('/').last;
 
   /// This function is called on resolve to
   /// get the data from the endpoint or cache.
   @override
   Data fromJson(String json) {
-    return _fromJson(json);
+    if (_fromJson != null) {
+      return _fromJson(json);
+    }
+
+    // Default implementation for direct usage
+    return json as Data;
+  }
+
+  @override
+  Future<String?> resolve() async {
+    try {
+      await super.hydratationCompleter.future;
+      final request = RepositoryHttpRequest(url: endpoint, method: method);
+      final response = await client.call(request: request);
+
+      /// If the endpoint returns a 200, add the data to the stream
+      /// and cache it
+      if (successfulCondition(response.statusCode, response.body)) {
+        final result = response.body;
+        return result;
+      } else {
+        BaseRepository.logger(
+          'Repository($name): Failed to resolve [$endpoint]. '
+          'status code: ${response.statusCode}',
+        );
+        final shouldThrow = await onErrorStatusCode(response.statusCode);
+        if (shouldThrow) {
+          throw UnexpectedStatusCodeException(
+            sent: request,
+            received: response,
+          );
+        } else {
+          return response.body;
+        }
+      }
+    } catch (exception) {
+      /// if the user is offline, the request will fail.
+      /// if [onSocketException] is not null, we call it.
+      /// if [onSocketException] is null, we rethrow the exception.
+      BaseRepository.logger(
+        'Repository($name): throws [${exception.runtimeType}].',
+        level: RepositoryLoggingLevel.warning,
+      );
+
+      // if (exception is Exception) {
+      //   await onSocketException(exception);
+      // }
+
+      rethrow;
+    }
+  }
+
+  /// If the request fails with [SocketException], this callback will be called
+  /// with the repository itself as an argument.
+  /// This callback is useful for retrying the request.
+  FutureOr<void> onSocketException(Exception exception) {
+    throw exception;
+  }
+
+  /// Condition to determine if the endpoint returns a successful status code.
+  /// The default condition is `statusCode == 200 || statusCode == 201`.
+  /// If you want to change the condition, override this method.
+  bool successfulCondition(int statusCode, dynamic body) {
+    return statusCode == 200 || statusCode == 201;
+  }
+
+  /// Called when the endpoint returns an unsuccessful status code.
+  /// See also [successfulCondition].
+  /// Return if the error should be thrown.
+  Future<bool> onErrorStatusCode(int statusCode) async {
+    return true;
+  }
+
+  @override
+  String get key {
+    /// We combine the `tag` and the `endpoint` to create a unique key
+    /// and we use the md5 hash to create a unique but "shorter" key.
+    /// When the `tag` is null, we use only the `endpoint` as the key.
+
+    if (tag == null) {
+      return endpoint.toString();
+    } else {
+      return '$tag,$endpoint';
+    }
   }
 
   @override
   FutureOr<bool> shouldRetry(Exception exception) {
     return _shouldRetryCondition?.call(exception) ??
-        super.shouldRetry(exception);
+        exception is NetworkUnavailableException;
+  }
+
+  @override
+  Future<void> mutate(Data data) async {
+    if (_mutate != null) {
+      return _mutate(data);
+    } else {
+      throw UnimplementedError('Mutate method not implemented');
+    }
   }
 }